refactor(core): extract the tick driver from the runner
Move the per-tick pass out of tui/runner.rs into core/driver/: playback
tick + debounced flushes, OAuth refresh (was per-frame in the draw block),
Discord/MPRIS/window-title presence sync, the lyrics + cover-art
track-change detector, native-queue and decoded-source auto-advance,
session persistence, and the visualizer capture. The runner draws frames,
reads events, and calls Driver::tick with a TickEnv carrying the
frontend-geometry inputs. A frontend that stops ticking now shows a stale
playbar via App::playback_position_ms (2s guard) instead of freezing
silently. Pure per-tick decisions live in core/driver/plan.rs with
deterministic tests.

Also: YouTube downloads retry once through the embedded player clients
(web_embedded,tv_embedded) when the default clients 403 under YouTube's
PO-token enforcement, which currently blocks most label uploads.
